change type data

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -6,11 +6,11 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int n = 0;
+	size_t n = 0;
 
 	while (h != NULL)
 	{
-		printf("%i\n", h->n);
+		printf("%d\n", h->n);
 		h = h->next;
 		n++;
 	}
